数据agent/requirements.txt
# 医学OCR Agent 依赖包

# HTTP请求库
requests>=2.31.0

# 环境变量管理
python-dotenv>=1.0.0

# Excel文件处理
openpyxl>=3.1.0

# FastAPI Web框架
fastapi>=0.115.0

# ASGI服务器
uvicorn[standard]>=0.32.0

# 文件上传支持
python-multipart>=0.0.12

# Pydantic数据验证
pydantic>=2.10.0

# 图像处理（可选，用于OCR预处理）
Pillow>=10.0.0